backend/chatbot-api/requirements.txt
# Core
fastapi>=0.95.0
uvicorn>=0.21.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
python-dotenv>=1.0.0

# Database
sqlalchemy>=2.0.0
alembic>=1.10.0
psycopg2-binary>=2.9.5

# AI & ML
openai>=1.3.7
langchain>=0.0.300
langchain-openai>=0.0.5
weaviate-client>=3.25.0

# Utils
pydantic>=2.0.0
pydantic-settings>=2.0.0
requests>=2.28.0
python-socketio>=5.8.0
python-engineio>=4.5.0
redis>=4.5.0
minio>=7.1.0

# Testing
pytest>=7.3.0
pytest-asyncio>=0.21.0
httpx>=0.24.0
